chore: moved test file
