+++ packages/lib/services/plugins/utils/isCompatible/index.test.ts
@@ -1,7 +1,5 @@
 import { AppType } from '../../../../models/Setting';
-import PluginService from '../../PluginService';
 import isCompatible from '../isCompatible';
-import { PluginManifest } from '../types';
 
 describe('isCompatible', () => {
 	test.each([
@@ -93,19 +91,5 @@ describe('isCompatible', () => {
 		expect(mobileCompatible).toBe(shouldSupportMobile);
 		const desktopCompatible = isCompatible(appVersion, AppType.Desktop, fullManifest);
 		expect(desktopCompatible).toBe(shouldSupportDesktop);
-	});
-
-	it('should report a missing app_min_version field specifically', () => {
-		const service = PluginService.instance();
-		service.initialize('3.0.0', {}, null, { dispatch: () => {} });
-		const manifest = {
-			id: 'test.plugin',
-			name: 'Test Plugin',
-			version: '1.0.0',
-			// Missing app_min_version
-		};
-
-		const error = service.describeIncompatibility(manifest as unknown as PluginManifest);
-		expect(error).toContain('missing the required "app_min_version" field');
 	});
 });
